AINFRA-283 - Set `queue: mac` explicitly in pipeline
This will allow adopting an improved default pipeline upload step on the
CI infrastructure side of the setup. It also makes it clear what queue
is in use just by looking at the pipeline file.

Notice it does not introduce the standard `shared-pipeline-vars` file
because to use that we first need the default pipeline upload step in
the IaC configuration.